Move blender file packing logic into the BlenderFilePacker class
